% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{Dendrometer032014}
\alias{Dendrometer032014}
\title{Dendrometer032014}
\format{A data frame with 7084 rows and 16 variables}
\description{
A dataset containing Dendrometer Data from March 2014
}
\details{
\itemize{
  \item Order
  \item tag
  \item plot
  \item q20
  \item p5
  \item many more variables
  ...
}
}
\keyword{datasets}

