% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{functionalTraits}
\alias{functionalTraits}
\title{Functional Traits Overall}
\format{A data frame with 122 rows and 35 variables}
\usage{
data(functionalTraits)
}
\description{
A dataset containing Functional Traits Overall
}
\details{
\itemize{
  \item species
  \item Chlorophyll
  \item Toughness
  \item Thickness
  \item Many more variables
  ...
}
}
\keyword{datasets}

